add the changes that have long, long been on the server

+++ src/main/serverside/upload.pl
@@ -8,42 +8,59 @@ use DBD::mysql;
 print qq(Content-type: text/plain\n);
 print qq(\n);
 
-my $dbh=DBI->connect("DBI:mysql:database=emcmarket;host=localhost", "emcmarket", "NE6xYRe6UPwW3Ray", {AutoCommit => 0 });
 
 my $cgi=CGI->new;
 my $rows=0;
-if ($cgi->param("upload")) {
-	if ($cgi->param("clientversion") eq "") {
-	} else {
-		my $name=$cgi->param("name");
-		my $separator=":";
-		if ($cgi->param("clientversion") ne "") {
-			$separator="\\|";
-		}
-		my $sthdel=$dbh->prepare(qq(
-			delete from signs
-			where server=? and x=? and y=? and z=? and chooseposition=?
-		));
-		my $sthins=$dbh->prepare(qq(
-			insert into signs (server, x, y, z, amount, buy, sell, owner, item, resnumber, chooseposition, lastseen, uploader, todelete)
-			values (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-		));
-		foreach my $row (split("\n", $cgi->param("upload"))) {
-			my ($server, $x, $y, $z, $amount, $buy, $sell, $owner, $item, $resnumber, $choosepos, $lastseen, $todelete)=
-				split($separator, $row);
-			$choosepos="0" if $choosepos eq "";
-			$lastseen=time*1000 if $lastseen eq "";
-			if ($todelete eq "todelete") { $todelete = 1; }
-			else { $todelete=0; }
+my $dbh;
+
+if ($cgi->param("clientversion") ne "" && $cgi->param("upload")) {
+	$dbh=DBI->connect("DBI:mysql:database=emcmarket;host=localhost", "emcmarket", "NE6xYRe6UPwW3Ray", {AutoCommit => 0 });
+	my $name=$cgi->param("name");
+	exit unless $name =~ /^[a-zA-Z0-9]+$/;
+	$name=$&;
+	open(DEBUG, ">/tmp/upload.debug/".$name.".".time);
+	my $separator=":";
+	if ($cgi->param("clientversion") ne "") {
+		$separator="\\|";
+	}
+	my $sthsel=$dbh->prepare(qq(
+		select count(*) from signs
+		where server=? and x=? and y=? and z=? and chooseposition=? and lastseen > ?
+	));
+	my $sthdel=$dbh->prepare(qq(
+		delete from signs
+		where server=? and x=? and y=? and z=? and chooseposition=?
+	));
+	my $sthins=$dbh->prepare(qq(
+		insert into signs (server, x, y, z, amount, buy, sell, owner, item, resnumber, chooseposition, lastseen, uploader, todelete)
+		values (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
+	));
+	foreach my $row (split("\n", $cgi->param("upload"))) {
+		my ($server, $x, $y, $z, $amount, $buy, $sell, $owner, $item, $resnumber, $choosepos, $lastseen, $todelete)=
+			split($separator, $row);
+		$choosepos="-1" if $choosepos eq "";
+		$lastseen=time*1000 if $lastseen eq "";
+		if ($todelete eq "todelete") { $todelete = 1; }
+		else { $todelete=0; }
+		$sthsel->execute($server, $x, $y, $z, $choosepos, $lastseen);
+		if (((my $count)=$sthsel->fetchrow_array()) && $count>0) {
+			print DEBUG "ignore $row\n";
+		} else {
 			$sthdel->execute($server, $x, $y, $z, $choosepos);
 			$sthins->execute($server, $x, $y, $z, $amount, $buy, $sell, $owner, $item, $resnumber, $choosepos, $lastseen, $name, $todelete) or print "$row : ".$dbh->errstr."\n";
-			$rows++;
+			print DEBUG "updated $row\n";
 		}
-		$dbh->commit;
+		$rows++;
 	}
+	close DEBUG;
+	$sthsel->finish;
+	$sthins->finish;
+	$sthdel->finish;
+	$dbh->commit;
 	open(F, ">>/tmp/uploaders.dat");
-	print F scalar localtime, "  ",  $rows ," uploaded by ", $cgi->param("name"), " client ", $cgi->param("clientversion"), "\n";
+	print F scalar localtime, "  ",  $rows ," uploaded by ", scalar $cgi->param("name"), " client ", scalar $cgi->param("clientversion"), "\n";
 	close F;
+	$dbh->disconnect;
 }
 
 print "$rows rows uploaded.\n";
